Update music player component so that when the state of toggle changes, the music starts / stops
