+++ frontend/src/App.js
@@ -11,7 +11,6 @@ import { useEffect, useState } from "react";
 import { setCurrentScene } from "./redux-store/scene/sceneSlice";
 import Bookshelf from "./components/bookshelf/Bookshelf.component";
 import MusicPlayer from "./components/music-player/music-player.component";
-import AudioPlayer from "./components/music-player/new-music-player";
 
 function App() {
   const dispatch = useDispatch();
@@ -30,7 +29,7 @@ function App() {
 
   return (
     <div className="flex flex-col justify-center items-center overflow-hidden h-screen">
-      <AudioPlayer
+      <MusicPlayer
         currentSongIndex={currentSongIndex}
         setCurrentSongIndex={setCurrentSongIndex}
       />
