+++ frontend/src/components/music-player/music-player.component.js
@@ -1,3 +1,4 @@
+import { useState, useEffect } from "react";
 import mononoke from "../../music/mononoke.mp3";
 import kiki from "../../music/kiki.mp3";
 import pokemon from "../../music/pokemon.mp3";
@@ -5,24 +6,47 @@ import porco from "../../music/porco.mp3";
 import spirited from "../../music/spirited.mp3";
 import wind from "../../music/wind.mp3";
 import { useSelector } from "react-redux";
-import React, { useState, useEffect, useRef } from "react";
 
-export default function MusicPlayer() {
+function MusicPlayer({ currentSongIndex, setCurrentSongIndex }) {
+  const [audioElement, setAudioElement] = useState(null);
   const { toggle } = useSelector((state) => state.musicPlayer);
   const trackList = [mononoke, kiki, pokemon, porco, spirited, wind];
 
-  const [trackIndex, setTrackIndex] = useState(0);
-  const [trackProgress, setTrackProgress] = useState(0);
-  const [isPlaying, setIsPlaying] = useState(false);
-  const audioElement = new Audio();
+  useEffect(() => {
+    setAudioElement(new Audio());
+  }, []);
 
-  return (
-    <>
-      {toggle ? (
-        <audio src={mononoke} controls style={{ display: "none" }} />
-      ) : (
-        console.log("music off")
-      )}
-    </>
-  );
+  useEffect(() => {
+    if (audioElement) {
+      audioElement.src = trackList[currentSongIndex];
+      if (toggle === true) {
+        audioElement.play();
+      } else {
+        audioElement.pause();
+      }
+    }
+  }, [audioElement, currentSongIndex, trackList, toggle]);
+
+  useEffect(() => {
+    if (audioElement) {
+      audioElement.addEventListener("ended", handleSongEnded);
+      return () => {
+        audioElement.removeEventListener("ended", handleSongEnded);
+      };
+    }
+  }, [audioElement, handleSongEnded]);
+
+  function handleSongEnded() {
+    if (currentSongIndex < trackList.length - 1) {
+      setCurrentSongIndex(currentSongIndex + 1);
+    }
+  }
+
+  function pauseSong() {
+    audioElement.pause();
+  }
+
+  return null;
 }
+
+export default MusicPlayer;
